Consolidate mongo files
